replace sqlite ConnectionPool with a plain SqlBackend
The sqlite connection pool leaks connections. Even with a capacity
of 1, it will leak connections until the process runs out of file
descriptors and needs to be restarted.

I've been delaying fixing this for an eternity because I was lazy,
and I was hoping I would spend the time to fix the root bug
in the persist-sqlite code. But I haven't done that yet, and there's
no reason to suspect that I will.

Using a ConnectionPool with the sqlite backend doesn't make sense
*anyway*, even if it didn't leak connections, because it can error
with "resource locked" (or whatever it is) if you are using a
pool size greater than one.